AdHocFilters: Better match old handling of queryparam formatting for ad hoc filters
