+++ packages/scenes/src/variables/adhoc/AdHocFiltersVariableUrlSyncHandler.ts
@@ -1,6 +1,6 @@
 import { SceneObjectUrlSyncHandler, SceneObjectUrlValue, SceneObjectUrlValues } from '../../core/types';
-import { AdHocFiltersVariable, AdHocFilterWithLabels, isFilterComplete, isMultiValueOperator } from './AdHocFiltersVariable';
-import { escapeUrlPipeDelimiters, toUrlCommaDelimitedString, unescapeUrlDelimiters } from '../utils';
+import { AdHocFiltersVariable, AdHocFilterWithLabels, isMultiValueOperator } from './AdHocFiltersVariable';
+import { unescapeUrlDelimiters } from '../utils';
 
 export class AdHocFiltersVariableUrlSyncHandler implements SceneObjectUrlSyncHandler {
   public constructor(private _variable: AdHocFiltersVariable) {}
@@ -14,14 +14,7 @@ export class AdHocFiltersVariableUrlSyncHandler implements SceneObjectUrlSyncHan
   }
 
   public getUrlState(): SceneObjectUrlValues {
-    const filters = this._variable.state.filters;
-
-    if (filters.length === 0) {
-      return { [this.getKey()]: [''] };
-    }
-
-    const value = filters.filter(isFilterComplete).map((filter) => toArray(filter).map(escapeUrlPipeDelimiters).join('|'));
-    return { [this.getKey()]: value };
+    return { [this.getKey()]: this._variable.getValueForUrl()?.toString() ?? '' };
   }
 
   public updateFromUrl(values: SceneObjectUrlValues): void {
@@ -44,23 +37,6 @@ function deserializeUrlToFilters(value: SceneObjectUrlValue): AdHocFilterWithLab
 
   const filter = toFilter(value);
   return filter === null ? [] : [filter];
-}
-
-function toArray(filter: AdHocFilterWithLabels): string[] {
-  const result = [
-    toUrlCommaDelimitedString(filter.key, filter.keyLabel),
-    filter.operator,
-  ];
-  if (isMultiValueOperator(filter.operator)) {
-    // TODO remove expect-error when we're on the latest version of @grafana/data
-    // @ts-expect-error
-    filter.values.forEach((value, index) => {
-      result.push(toUrlCommaDelimitedString(value, filter.valueLabels?.[index]));
-    });
-  } else {
-    result.push(toUrlCommaDelimitedString(filter.value, filter.valueLabels?.[0]));
-  }
-  return result;
 }
 
 function toFilter(urlValue: string | number | boolean | undefined | null): AdHocFilterWithLabels | null {
